Using a shader to draw the border
We used to draw the border by setting different features of a
PaintFlags. This CL moves all those logic to a SkSL command which draws
the border given some uniforms.

Bug: 388568973
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6179759
Commit-Queue: Baran Erfani <baranerf@google.com>
Cr-Commit-Position: refs/heads/main@{#1408006}

+++ chrome/browser/glic/border_view_browsertest.cc
@@ -84,7 +84,7 @@ class BorderViewBrowserTest : public InteractiveBrowserTest {
   static SkBitmap ConstructExpectedBitmap(const gfx::Size& size,
                                           SkColor border_color,
                                           SkColor center_color,
-                                          int border_width,
+                                          float border_width,
                                           float alpha) {
     SkBitmap bitmap;
     SkImageInfo info =
@@ -189,7 +189,7 @@ IN_PROC_BROWSER_TEST_F(BorderViewBrowserTest, AnimationSteps) {
     SkBitmap expected_bitmap = ConstructExpectedBitmap(
         capture_rect.size(),
         /*border_color=*/BorderColor(),
-        /*center_color=*/kBlack, /*border_width=*/2, /*alpha=*/0.f);
+        /*center_color=*/kBlack, /*border_width=*/2.f, /*alpha=*/0.f);
 
     EXPECT_TRUE(cc::MatchesBitmap(
         actual_bitmap, expected_bitmap,
@@ -211,7 +211,7 @@ IN_PROC_BROWSER_TEST_F(BorderViewBrowserTest, AnimationSteps) {
     // The border width is calculated as:
     // `kBorderWidthMin` + ((`kBorderWidthMax` - `kBorderWidthMin`) *
     // `progress`).
-    int border_width = 2 + (8 * progress);
+    float border_width = 2 + (8 * progress);
     SkBitmap expected_bitmap = ConstructExpectedBitmap(
         capture_rect.size(),
         /*border_color=*/BorderColor(),
@@ -232,7 +232,7 @@ IN_PROC_BROWSER_TEST_F(BorderViewBrowserTest, AnimationSteps) {
     SkBitmap expected_bitmap =
         ConstructExpectedBitmap(capture_rect.size(),
                                 /*border_color=*/BorderColor(),
-                                /*center_color=*/kBlack, /*border_width=*/10,
+                                /*center_color=*/kBlack, /*border_width=*/10.f,
                                 /*alpha=*/1.f);
 
     EXPECT_TRUE(cc::MatchesBitmap(
@@ -249,7 +249,7 @@ IN_PROC_BROWSER_TEST_F(BorderViewBrowserTest, AnimationSteps) {
     SkBitmap expected_bitmap = ConstructExpectedBitmap(
         capture_rect.size(),
         /*border_color=*/BorderColor(),
-        /*center_color=*/kBlack, /*border_width=*/10, /*alpha=*/1.f);
+        /*center_color=*/kBlack, /*border_width=*/10.f, /*alpha=*/1.f);
 
     EXPECT_TRUE(cc::MatchesBitmap(
         actual_bitmap, expected_bitmap,
@@ -298,7 +298,7 @@ IN_PROC_BROWSER_TEST_F(BorderViewBrowserTest, FocusedTabChange) {
     SkBitmap expected_bitmap = ConstructExpectedBitmap(
         capture_rect.size(),
         /*border_color=*/BorderColor(),
-        /*center_color=*/kBlack, /*border_width=*/2, /*alpha=*/0.f);
+        /*center_color=*/kBlack, /*border_width=*/2.f, /*alpha=*/0.f);
 
     EXPECT_TRUE(cc::MatchesBitmap(
         actual_bitmap, expected_bitmap,
@@ -312,7 +312,7 @@ IN_PROC_BROWSER_TEST_F(BorderViewBrowserTest, FocusedTabChange) {
     SkBitmap actual_bitmap = PaintBorder(border);
 
     float progress = sin(0.125 * M_PI);
-    int border_width = 2 + (8 * progress);
+    float border_width = 2 + (8 * progress);
     SkBitmap expected_bitmap = ConstructExpectedBitmap(
         capture_rect.size(),
         /*border_color=*/BorderColor(),
@@ -340,7 +340,7 @@ IN_PROC_BROWSER_TEST_F(BorderViewBrowserTest, FocusedTabChange) {
     SkBitmap expected_bitmap =
         ConstructExpectedBitmap(capture_rect.size(),
                                 /*border_color=*/BorderColor(),
-                                /*center_color=*/kBlack, /*border_width=*/2,
+                                /*center_color=*/kBlack, /*border_width=*/2.f,
                                 /*alpha=*/0.f);
 
     EXPECT_TRUE(cc::MatchesBitmap(
@@ -355,7 +355,7 @@ IN_PROC_BROWSER_TEST_F(BorderViewBrowserTest, FocusedTabChange) {
     SkBitmap actual_bitmap = PaintBorder(border);
 
     float progress = sin(0.125 * M_PI);
-    int border_width = 2 + (8 * progress);
+    float border_width = 2 + (8 * progress);
     SkBitmap expected_bitmap = ConstructExpectedBitmap(
         capture_rect.size(),
         /*border_color=*/BorderColor(),
@@ -397,7 +397,7 @@ IN_PROC_BROWSER_TEST_F(BorderViewBrowserTest, FocusedWindowChange) {
     SkBitmap expected_bitmap = ConstructExpectedBitmap(
         capture_rect.size(),
         /*border_color=*/BorderColor(),
-        /*center_color=*/kBlack, /*border_width=*/2, /*alpha=*/0.f);
+        /*center_color=*/kBlack, /*border_width=*/2.f, /*alpha=*/0.f);
 
     EXPECT_TRUE(cc::MatchesBitmap(
         actual_bitmap, expected_bitmap,
@@ -425,7 +425,7 @@ IN_PROC_BROWSER_TEST_F(BorderViewBrowserTest, FocusedWindowChange) {
     SkBitmap expected_bitmap = ConstructExpectedBitmap(
         capture_rect.size(),
         /*border_color=*/BorderColor(),
-        /*center_color=*/kBlack, /*border_width=*/2, /*alpha=*/0.f);
+        /*center_color=*/kBlack, /*border_width=*/2.f, /*alpha=*/0.f);
 
     EXPECT_TRUE(cc::MatchesBitmap(
         actual_bitmap, expected_bitmap,
@@ -437,7 +437,7 @@ IN_PROC_BROWSER_TEST_F(BorderViewBrowserTest, FocusedWindowChange) {
     SkBitmap actual_bitmap = PaintBorder(new_border);
 
     float progress = sin(0.25 * M_PI);
-    int border_width = 2 + (8 * progress);
+    float border_width = 2 + (8 * progress);
     SkBitmap expected_bitmap = ConstructExpectedBitmap(
         capture_rect.size(),
         /*border_color=*/BorderColor(),
